refactor(ImageUploader): move the compoent to independent file
